feat: 지역에 가게 추가하는 API

+++ src/main/java/javalab/umc7th_mission/repository/FoodCategoryRepository.java
@@ -3,5 +3,5 @@ package javalab.umc7th_mission.repository;
 import javalab.umc7th_mission.domain.FoodCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FoodCategoryRepository extends JpaRepository<FoodCategory, Long> {
+public interface FoodCategoryRepository extends JpaRepository<FoodCategory, Integer> {
 }
