+++ src/main/java/javalab/umc7th_mission/repository/UserFoodCategoryRepository.java
@@ -3,6 +3,6 @@ package javalab.umc7th_mission.repository;
 import javalab.umc7th_mission.domain.mapping.UserFoodCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserFoodCategoryRepository extends JpaRepository<UserFoodCategory, Long> {
+public interface UserFoodCategoryRepository extends JpaRepository<UserFoodCategory, Integer> {
 
 }
